fix: dont clobber static config with overrides (#50)
## what

- Updates `runtime_overrides` for a given stack to be included in the
deep merge instead of TF's simple merge.


## why

- Short version: Properly merges runtime overrides on top of static
StackConfig values
- Long version: I found this bug when doing some diligent integration
testing in our mp-infra repo. Prior to this fix, if we included
`runtime_overrides` with only one value in the overrides, it would
overwrite the static StackConfig `stack_settings` for that stack and
result in only including the runtime_overrides + the variable defaults.
Not good 💥
- This is exactly what @westonplatter was working to avoid when he was
pushing me in this thread:
https://github.com/masterpointio/terraform-spacelift-automation/pull/44#discussion_r1936465439.
He added a test for this, but it was with `labels` and he was checking a
common label... sadly that didn't catch this bug even though he was spot
on that there was bug here!!! Luckily, integration testing surfaced this
and then unit testing enabled me to quickly reproduce in the module. Now
we have a solid fix 🎉

## references

- Follow up to #44 and #48


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Refactor**
- Enhanced the configuration merging process to now apply runtime
overrides after base and stack-specific settings, offering greater
flexibility and dynamic control.
- **Tests**
- Expanded test coverage with additional validations and refined naming
to ensure consistent behavior and reliability of the configuration
adjustments.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->