Make `terraform_version` optional
This module supports a `terraform_version` input variable that is used
to create `spacelift_stack` resources.

`terraform_version` is considered optional by the Spacelift provider.

This PR proposes to change the variable type to `optional(string)`.

This PR also proposes to set the default to `null` to better match the
Spacelift provider. The default is currently `1.7.1`, which is outdated
for OpenTofu but also too high for `TERRAFORM_FOSS`.

There doesn't seem to be any need for a validator on the variable. The
only restrictions on `terraform_version` are related to the nested
`terragrunt` schema, which is not supported by this module currently.

https://github.com/spacelift-io/terraform-provider-spacelift