feat: set default `terraform_version = null` (#73)
## Description

Discussed with @Gowiem in Slack 2025-06-13.

This module supports a `terraform_version` input variable that is used
to create `spacelift_stack` resources.


https://github.com/masterpointio/terraform-spacelift-automation/blob/af2a2dea6209d00416059ee4c9a38aa41d961445/main.tf#L364

`terraform_version` is considered optional by the Spacelift provider
(see source code
[here](https://github.com/spacelift-io/terraform-provider-spacelift/blob/97e4ac0710053b19868456df0c078ad9232c4db4/spacelift/resource_stack.go#L608-L613),
[here](https://github.com/spacelift-io/terraform-provider-spacelift/blob/97e4ac0710053b19868456df0c078ad9232c4db4/spacelift/resource_stack.go#L1077-L1088)
and
[here](https://github.com/spacelift-io/terraform-provider-spacelift/blob/97e4ac0710053b19868456df0c078ad9232c4db4/spacelift/resource_stack.go#L1037-L1039)).

## Changes

This PR proposes to change the variable default to `null` to better
match the Spacelift provider. The default is currently `1.7.1`, which is
outdated for OpenTofu but also too high for `TERRAFORM_FOSS`.

There doesn't seem to be any need for a validator on the variable. The
only restrictions on `terraform_version` are related to the [nested
`terragrunt`
schema](https://search.opentofu.org/provider/spacelift-io/spacelift/latest/docs/resources/stack#nested-schema-for-terragrunt),
which is not supported by this module currently.

## Related

- [OpenTofu Registry: Provider - spacelift-io/spacelift -
`spacelift_stack`](https://search.opentofu.org/provider/spacelift-io/spacelift/latest/docs/resources/stack)
-
https://github.com/masterpointio/terraform-spacelift-automation/pull/54

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Documentation**
- Updated descriptions for the Terraform version input to clarify that
it now supports both OpenTofu and Terraform, and defaults to the latest
available version if not specified.
- **New Features**
- The Terraform version input is now optional and will automatically use
the latest version if not explicitly set.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Matt Gowie <matt@masterpoint.io>
Co-authored-by: yangci <yangciouou@gmail.com>